refactor : 게시글 수정 API 리팩토링
리팩 토링 내용
- 기존 게시글 수정시 이미지를 필수로 보내야했는데 이제는 이미지의 데이터를 넘기지 않으면 이미지 변경은 하지 않도록 리팩토링

+++ src/main/java/com/team/RecipeRadar/domain/post/application/PostServiceImpl.java
@@ -134,18 +134,19 @@ public class PostServiceImpl implements PostService {
      * 게시글을 업데이트 하기 위한 로직
      */
     @Override
-    public void update(UserUpdateRequest userUpdateRequest,String loginId,MultipartFile file) {
-        Long postId = userUpdateRequest.getPostId();
+    public void update(Long postId,UserUpdateRequest userUpdateRequest,String loginId,MultipartFile file) {
 
         Post post = postRepository.findById(postId).orElseThrow(() -> new NoSuchElementException("해당 게시물을 찾을 수 없습니다."));
         if(!post.getMember().getLoginId().equals(loginId)) throw new AccessDeniedException("작성자만 삭제 가능합니다.");
 
         UploadFile uploadFile = imgRepository.getOriginalFileName(post.getId());
-        if (!uploadFile.getOriginFileName().equals(file.getOriginalFilename())) {       // 원본파일명이 다를경우에만 s3에 기존 사진을 삭제 후 새롭게 저장
-            s3UploadService.deleteFile(uploadFile.getStoreFileName());
-            String storedFileName = s3UploadService.uploadFile(file);
-            uploadFile.update(storedFileName, file.getOriginalFilename());
-            imgRepository.save(uploadFile);
+        if(file!=null) {
+            if (!uploadFile.getOriginFileName().equals(file.getOriginalFilename())) {       // 원본파일명이 다를경우에만 s3에 기존 사진을 삭제 후 새롭게 저장
+                s3UploadService.deleteFile(uploadFile.getStoreFileName());
+                String storedFileName = s3UploadService.uploadFile(file);
+                uploadFile.update(storedFileName, file.getOriginalFilename());
+                imgRepository.save(uploadFile);
+            }
         }
 
         post.update(userUpdateRequest.getPostTitle(), userUpdateRequest.getPostContent(), userUpdateRequest.getPostServing(),
